Fix for "net -n" option to properly deal with an invalid argument
The help/man page of the "net" command suggests that "-n" option can
accept two kinds of argument: PID or task_struct pointer.  However,
the "net -n" command accepts an invalid argument and shows the
namespace of the current context silently.  For example:

  crash> net -n 1000000000
     NET_DEVICE     NAME   IP ADDRESS(ES)
  ffff949dc11d7000  lo     127.0.0.1
  ffff949dcc01c000  eno49  192.168.122.17

With the patch, emit an error expectedly.

  crash> net -n 1000000000
  net: invalid task or pid value: 1000000000

Reported-by: Buland Kumar Singh <bsingh@redhat.com>
Signed-off-by: Lianbo Jiang <lijiang@redhat.com>