Fix "vm -M" option to properly deal with an invalid argument
The "vm -M" option can accept an invalid address and print the virtual
memory data of a task without an error like this:

   crash> vm -M 0xdeadbeef
   PID: 92960    TASK: ffff99157976cc80  CPU: 0    COMMAND: "crash"
          MM               PGD          RSS    TOTAL_VM
   ffff991573bfdf00  ffff9915857f2000  449020k  2427076k
         VMA           START       END     FLAGS FILE
   ffff99158718d1c8     400000     4de000 8000071 /home/crash/crash
   ...

The reasons are
- htoll() only converts a hexadecimal string to an unsigned long long
   value and does not evaluate whether it's a valid kernel virtual
   address or not, and
- The specified value is used only when the task's mm_struct is NULL.

Also, this behavior is not described enough in its help text, so it's
confusing for users.

Let's add a check on the converted value regardless of the task's
mm_struct and add a description of the behavior to its help text.

With the patch:
   crash> vm -M 0xdeadbeef
   vm: invalid mm_struct address: 0xdeadbeef

Reported-by: Buland Kumar Singh <bsingh@redhat.com>
Signed-off-by: Lianbo Jiang <lijiang@redhat.com>
[ kh: rewrote commit message ]
Signed-off-by: Kazuhito Hagio <k-hagio-ab@nec.com>