Fix "kmem -v" option displaying no regions on Linux 6.3 and later
Kernel commit 869176a09606 ("mm/vmalloc.c: add flags to mark vm_map_ram
area"), which is contained in Linux 6.3 and later, added "flags" member
to struct vmap_area.  This was the revival of the "flags" member as
kernel commit 688fcbfc06e4 had eliminated it before.

As a result, crash started to use the old procedure using the member and
displays no vmalloc'd regions, because it does not have the same flag
value as the old one.

  crash> kmem -v
     VMAP_AREA         VM_STRUCT                 ADDRESS RANGE                SIZE
  crash>

To fix this, also check if vmap_area.purge_list exists, which was
introduced with the flags and removed later, to determine that the flags
member is the old one.

Related vmap_area history:
 v2.6.28 db64fe02258f introduced vmap_area with flags and purge_list
 v5.4    688fcbfc06e4 removed flags
 v5.11   96e2db456135 removed purge_list
 v6.3    869176a09606 added flags again

Signed-off-by: Kazuhito Hagio <k-hagio-ab@nec.com>