Fix invalid structure size error during crash startup on ppc64
The crash utility will fail to start session on ppc64 with the following
error:

  # crash vmlinux vmcore -s

  crash: invalid structure size: note_buf
         FILE: diskdump.c  LINE: 121  FUNCTION: have_crash_notes()

  [./crash] error trace: 101859ac => 10291798 => 10291450 => 10266038

    10266038: SIZE_verify+156
    10291450: have_crash_notes+308
    10291798: map_cpus_to_prstatus_kdump_cmprs+448
    101859ac: task_init+11980

The reason is that the size of note_buf is not initialized before using
SIZE(note_buf) in the have_crash_notes() on some architectures including
ppc64.  Let's initialize it in task_init() to fix this issue.

Fixes: db8c030857b4 ("diskdump/netdump: fix segmentation fault caused by failure of stopping CPUs")
Signed-off-by: Lianbo Jiang <lijiang@redhat.com>