vmware: Improve output when we fail to read vmware 'vmsn' file
Today if crash fails to read some structure in a vmware 'vmsn' file,
it will throw an "No such file or directory" message.  Such a generic
message does not give any clue as to the problem, but instead sounds
like the file may not exist when it does, for example:
  $ crash ./vmcore.vmsn  ./vmlinux

  crash 8.0.3
  ...

  crash: vmw: Failed to read './vmcore.vmsn': [Error 2] No such file or directory

  crash: ./vmcore.vmsn: initialization failed

  $ ls -l ./vmcore.vmsn
  -rwxrwxrwx. 7 myuser mygroup 12128999 Jul  4 07:21 ./vmcore.vmsn

Improve the above error message so we at least know which portion
of the file crash had difficulty reading.  After this patch, the
above error looks like:
  crash: vmw: Failed to read 'cptgroupdesc' from file './vmcore.vmsn': [Error 2] No such file or directory

Signed-off-by: Dave Wysochanski <dwysocha@redhat.com>