Revert "Fix "kmem -s|-S" not working properly on RHEL8.6 and later"
This reverts commit 9253b40a0ecb2d365f89f0a5ebc28a01735c1d24.

The commit 9253b40a0ecb only handles the current issue on x86_64/x86
architectures. Furthermore the freelist_ptr_bswap_x86() depends on
disassembling a static symbol which might not be available, depending on
how the compiler decides to optimize the code, that is to say, the
compiler might generate different code eventually.

More importantly, a subsequent patch can cover the current issue on
various architectures. Given that, revert the commit.

Signed-off-by: Lianbo Jiang <lijiang@redhat.com>