In verify_version() don't require specific syment type values for linux_banner symbol to get its address
verify_version() in kernel.c gets a struct syment for linux_banner using
symbol_search() and uses the value member of the result as the address
of linux_banner in some cases based on the type member's value in the
same struct syment. A small number of coredumps with an unhandled type
('B' or 'b') for linux_banner result in the address of linux_banner
being loaded from the actual linux_banner data. This fails because the
first ulong of the linux_banner ASCII text is treated as a dumped kernel
address and attempting to access that in the core fails.

Based on a suggestion from Kazu, continue to get the struct syment for
linux_banner using symbol_search(). Also use get_symbol_type() for
linux_banner and use the result of that to decide where to get the
linux_banner address from, disregarding the syment type member. If
get_symbol_type() reports a TYPE_CODE_ARRAY (and by default with a
warning) use the syment value member as the linux_banner address. If
get_symbol_type() reports a TYPE_CODE_PTR read the address of
linux_banner using get_symbol_data().

The else block doesn't strictly require braced content for a single
switch statement but braces are included to match style of locally
similar cases.

Signed-off-by: David Mair <dmair@suse.com>