memory_driver: Ensure PWD points to the current directory
Building crash.ko broke in commit 74ac92971241 ("Support for multiple
jobs to build crash"), as PWD won't be updated on recursive calls to
'make' and will still point to the upper directory.

This leads to the 'all' target trying to build modules in the top level
directory -- where there are none.

Fix that by updating PWD to the current directory.

Fixes: 74ac92971241 ("Support for multiple jobs to build crash")
Reported-by: Lianbo Jiang <lijiang@redhat.com>
Signed-off-by: Mathias Krause <minipli@grsecurity.net>