LoongArch64: Make the crash tool successfully enter the crash command line
1. Add loongarch64_init() implementation, do all necessary machine-specific setup,
which will be called multiple times during initialization.

2. Add the implementation of the vtop command, which is used to convert a
virtual address to a physical address. When entering the crash command line,
the corresponding symbols in the kernel will be read, and at the same time,
the conversion of virtual and real addresses will also be used, so the vtop
command is a prerequisite for entering the crash command line.

3. Add loongarch64_get_smp_cpus() implementation, get the number of online cpus.

4. Add loongarch64_get_page_size() implementation, get page size.

5. Add to get processor speed. Obtain the processor speed from the kernel
symbol "cpu_clock_freq".

6. Add loongarch64_verify_symbol() implementation, accept or reject a symbol from
the kernel namelist.

With this patch, we can enter crash command line. Tested on Loongson-3C5000 platform.

For help, type "help".
Type "apropos word" to search for commands related to "word"...

      KERNEL: /usr/lib/debug/lib/modules/5.10.0-60.103.0.130.oe2203.loongarch64/vmlinux
    DUMPFILE: /proc/kcore
        CPUS: 16
        DATE: Mon Aug 21 14:33:19 CST 2023
      UPTIME: 05:01:34
LOAD AVERAGE: 0.43, 0.11, 0.17
       TASKS: 265
    NODENAME: localhost.localdomain
     RELEASE: 5.10.0-60.103.0.130.oe2203.loongarch64
     VERSION: #1 SMP Fri Jul 21 12:48:08 UTC 2023
     MACHINE: loongarch64  (2200 Mhz)
      MEMORY: 64 GB
         PID: 114499
     COMMAND: "crash"
        TASK: 900000009676ff00  [THREAD_INFO: 90000000981a8000]
         CPU: 12
       STATE: TASK_RUNNING (ACTIVE)

Co-developed-by: Youling Tang <tangyouling@loongson.cn>
Signed-off-by: Youling Tang <tangyouling@loongson.cn>
Signed-off-by: Ming Wang <wangming01@loongson.cn>