Optionally use orc_header to detect orc version
In Linux kernel commit b9f174c811e3ae4ae8959dc57e6adb9990e913f4, a new
.orc_header section was introduced.  This section holds a version
identifier for the ORC types in use.

This patch was tested by compiling crash and using it to read coredump
of kernel.  A temporary print statement was added to check that the
if conditions specified are working fine.