Adding the zram decompression algorithm "lzo-rle" to support kernel versions >= 5.1.
Port the improved decompression method for "lzo" in the kernel to support decompression of "lzorle". After the 5.1 version of the kernel, the default compression algorithm for zram was changed from "lzo" to "lzo-rle". the crash utility only supports decompression for "lzo", when parsing vmcore files that utilize zram compression, such as when using the gcore command to detach process core dump files, parsing cannot be completed successfully.

before:
	crash> gcore -v 0 1
	gcore: WARNING: only the lzo compressor is supported
	gcore: WARNING: only the lzo compressor is supported
	gcore: WARNING: only the lzo compressor is supported
	gcore: WARNING: only the lzo compressor is supported
after:
	crash> gcore -v 0 1
	Saved core.1.init