Prevent double-free strbuf when nsyms is 0
When nsyms is 0, 'last' will be equal to 'first', and strbuf will not
get a new memory buf. But strbuf is not NULL now, because in the last
loop, strbuf may get a memory buf allocated by 'calloc', which is only
freed in the loop end but not set to NULL. If we don't set strbuf to
NULL in such a case, strbuf will be double freed in the following
FREEBUF(strbuf) sentence.

Signed-off-by: Wang Guanghui <wanggh1@sugon.com>