Fix for "help -r" segfault in case of ramdump
When the ELF note does not contain CPU registers, attempting to retrieve
online CPU registers will cause a crash. This is likely to happen for
ramdump cases as follows, where the ELF header is created by crash at
runtime.

  crash> help -r
  ...
   WARNING: cpu 7: cannot find NT_PRSTATUS note
         KERNEL: vmlinux  [TAINTED]
      DUMPFILES: /var/tmp/ramdump_elf_RabECD [temporary ELF header]
                 DDRCS0_0.BIN
                 ...
                 DDRCS2_2.BIN
           CPUS: 8
  ...

With the patch:
  crash> help -r
  ...
  CPU 6:
  help: registers not collected for cpu 6
  ...

Lets add a sanity check to avoid the current issue.

Signed-off-by: Guanyou.Chen <chenguanyou@xiaomi.com>