x86_64: Mark #VC stack unavailable when CONFIG_AMD_MEM_ENCRYPT is not set
When 'CONFIG_AMD_MEM_ENCRYPT' is not set, the IDT handler for #VC is not
registered, but the address of the #VC IST stack is still set in
tss_setup_ist(). Therefore, the name of the associated exception stack
is "UNKNOWN" instead of "VC". Although the base of the exception stack
is not zero and is available, it is not accessible, which will cause the
backtrace to fail when attempting to access the #VC stack. To fix this,
remove the name check.

Signe-off-by: Hou Wenlong <houwenlong.hwl@antgroup.com>